Scale Measurement to default scaling factor
Before adding full scaling support this change scales the results to the
default scaling level. That is ±2g for the accelerometer and just
multiply by 150 for the magnetometer.

I have renamed `UnscaledMeasurement` to `Measurement` to make it clear
that these are now scaled.

The scaling factor to use for the accelerometer is listed in Table 14.
Operating Mode Selection in the datasheet in the last column. The 150
scaling value for the magnetometer is a little more obscure. It is
defined [here in the
microbit-dal](https://github.com/lancaster-university/microbit-dal/blob/master/inc/drivers/LSM303Magnetometer.h#L38)
but with very little context. The [commit that introduced
it](https://github.com/lancaster-university/microbit-dal/commit/f6bee0940eaf8c83529d51d6c6c2f45cfc90c266)
is equaly vague (although it does imply that the value has been determined
empirically). I have seen similar scaling factors used in other LSM303
drivers but have not been able to find an explanation or any reference
in the datasheet.